inserting a node at a given index

+++ 0x17-doubly_linked_lists/7-insert_dnodeint.c
@@ -12,7 +12,7 @@
 dlistint_t *insert_dnodeint_at_index(dlistint_t **h, unsigned int idx, int n)
 {
 	dlistint_t *new, *end;
-	unsigned int k = 0;
+	size_t int k = 0;
 
 	new = malloc(sizeof(dlistint_t));
 	if (new == NULL)
